backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
faster-whisper==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
pydantic==2.10.4
websockets==14.1
aiofiles==24.1.0
PyJWT==2.9.0
bcrypt==4.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
